HW-7: Object model refactored: ObjectListSection renamed to TitledListSection and it's  now can keep for each title a list of text fields with dates, ContactType Enum added,  Section now empty. to all model classes - added equals and hashCode. ResumeTestDate, AbstractStorage, ListStorage refactored.

+++ src/com/vkopendoh/webapp/storage/ListStorage.java
@@ -59,7 +59,7 @@ public class ListStorage extends AbstractStorage<Integer> {
 
     @Override
     protected void removeByKey(Integer searchKey) {
-        int index = searchKey.intValue();
+        int index = searchKey;
         storage.remove(index);
 
     }
